+++ projects/nuxtjs-no-slash/plugins/$path.ts
@@ -10,6 +10,9 @@ import { Query as Query2 } from '../pages/blog/_slug'
 
 // prettier-ignore
 export const pagesPath = {
+  _ignore: {
+    $url: (url?: { hash?: string }) => ({ path: '/.ignore', hash: url?.hash })
+  },
   _a: (a: string | number) => ({
     b: {
       _c: (c: string | number) => ({
@@ -40,9 +43,27 @@ export const pagesPath = {
 export type PagesPath = typeof pagesPath
 
 // prettier-ignore
+export const staticPath = {
+  aa_json: '/aa.json',
+  bb: {
+    _ignore: '/bb/.ignore',
+    cc_png: '/bb/cc.png'
+  },
+  duplicate_json_0: {
+    sample_json: '/duplicate-json/sample.json'
+  },
+  duplicate_json_1: '/duplicate.json',
+  duplicate_json_2: '/duplicate_json'
+} as const
+
+// prettier-ignore
+export type StaticPath = typeof staticPath
+
+// prettier-ignore
 declare module 'vue/types/vue' {
   interface Vue {
     $pagesPath: PagesPath
+    $staticPath: StaticPath
   }
 }
 
@@ -50,10 +71,12 @@ declare module 'vue/types/vue' {
 declare module '@nuxt/types' {
   interface NuxtAppOptions {
     $pagesPath: PagesPath
+    $staticPath: StaticPath
   }
 
   interface Context {
     $pagesPath: PagesPath
+    $staticPath: StaticPath
   }
 }
 
@@ -61,12 +84,14 @@ declare module '@nuxt/types' {
 declare module 'vuex/types/index' {
   interface Store<S> {
     $pagesPath: PagesPath
+    $staticPath: StaticPath
   }
 }
 
 // prettier-ignore
 const pathPlugin: Plugin = (_, inject) => {
   inject('pagesPath', pagesPath)
+  inject('staticPath', staticPath)
 }
 
 // prettier-ignore