+++ src/createSapperTemplate.ts
@@ -1,5 +1,6 @@
 import fs from 'fs'
 import path from 'path'
+import { createIg, isIgnored } from './isIgnored'
 import { parseQueryFromTS } from './parseQueryFromTS'
 import { replaceWithUnderscore } from './replaceWithUnderscore'
 
@@ -54,11 +55,17 @@ const parseQueryFromSvelte = (file: string, suffix: number) => {
   }
 }
 
-export default (input: string, trailingSlash = false) => {
+export default (
+  input: string,
+  output: string,
+  ignorePath: string | undefined,
+  trailingSlash = false
+) => {
+  const ig = createIg(ignorePath)
   const imports: string[] = []
   const getImportName = (file: string) => {
     const result = path.extname(file).startsWith('.ts')
-      ? parseQueryFromTS(input, file, imports.length)
+      ? parseQueryFromTS(output, file, imports.length)
       : parseQueryFromSvelte(file, imports.length)
 
     if (result) {
@@ -79,7 +86,13 @@ export default (input: string, trailingSlash = false) => {
 
     const files = fs
       .readdirSync(targetDir)
-      .filter(file => !file.startsWith('_') && !/\.s?css(\.d\.ts)?$/.test(file))
+      .filter(
+        file =>
+          !file.startsWith('_') &&
+          !/\.s?css$/.test(file) &&
+          !file.endsWith('.d.ts') &&
+          !isIgnored(ig, ignorePath, targetDir, file)
+      )
     const props: string[] = [
       ...files,
       ...files