+++ src/createStaticTemplate.ts
@@ -1,13 +1,10 @@
 import fs from 'fs'
 import path from 'path'
+import { createIg, isIgnored } from './isIgnored'
 import { replaceWithUnderscore } from './replaceWithUnderscore'
 
-const normalizeBasePath = (basepath: string | undefined): string => {
-  if (typeof basepath === 'string') return basepath.replace(/\/+$/, '')
-  return ''
-}
-
-export default (input: string, basepath: string | undefined) => {
+export default (input: string, basepath: string | undefined, ignorePath: string | undefined) => {
+  const ig = createIg(ignorePath)
   const createPublicString = (targetDir: string, indent: string, url: string, text: string) => {
     indent += '  '
 
@@ -21,6 +18,9 @@ export default (input: string, basepath: string | undefined) => {
       .map((file, i) => {
         const newUrl = `${url}/${file}`
         const target = path.posix.join(targetDir, file)
+
+        if (isIgnored(ig, ignorePath, targetDir, file)) return ''
+
         const replacedFile = replacedFiles[i]
         const valFn = `${indent}${
           duplicatedInfo[replacedFile].length > 1
@@ -47,7 +47,7 @@ export default (input: string, basepath: string | undefined) => {
   const text = createPublicString(
     input,
     '',
-    normalizeBasePath(basepath),
+    typeof basepath === 'string' ? basepath.replace(/\/+$/, '') : '',
     `{\n<% props %>\n} as const`
   )