+++ src/getConfig.ts
@@ -1,12 +1,13 @@
 import fs from 'fs'
-import path from 'path'
 import type { NextConfig } from 'next/dist/server/config'
+import path from 'path'
 
 export type Config = {
   type: 'nextjs' | 'nuxtjs' | 'sapper'
   input: string
-  staticDir?: string
+  staticDir: string | undefined
   output: string
+  ignorePath: string | undefined
   trailingSlash?: boolean
   basepath?: string
 }
@@ -18,8 +19,14 @@ const getFrameworkType = (dir: string) => {
   return deps.sapper ? 'sapper' : deps.nuxt ? 'nuxtjs' : 'nextjs'
 }
 
-export default async (enableStatic: boolean, dir = process.cwd()): Promise<Config> => {
+export default async (
+  enableStatic: boolean,
+  output: string | undefined,
+  igPath: string | undefined,
+  dir = process.cwd()
+): Promise<Config> => {
   const type = getFrameworkType(dir)
+  const ignorePath = igPath && path.join(dir, igPath)
 
   if (type === 'nextjs') {
     let config: NextConfig
@@ -39,8 +46,11 @@ export default async (enableStatic: boolean, dir = process.cwd()): Promise<Confi
     }
 
     const srcDir = fs.existsSync(path.posix.join(dir, 'pages')) ? dir : path.posix.join(dir, 'src')
-    const utilsPath = path.join(srcDir, 'utils')
-    const output = fs.existsSync(utilsPath) ? utilsPath : path.join(srcDir, 'lib')
+
+    if (!output) {
+      const utilsPath = path.join(srcDir, 'utils')
+      output = fs.existsSync(utilsPath) ? utilsPath : path.join(srcDir, 'lib')
+    }
 
     if (!fs.existsSync(output)) fs.mkdirSync(output)
 
@@ -49,6 +59,7 @@ export default async (enableStatic: boolean, dir = process.cwd()): Promise<Confi
       input: path.posix.join(srcDir, 'pages'),
       staticDir: enableStatic ? path.posix.join(dir, 'public') : undefined,
       output,
+      ignorePath,
       basepath: config.basePath
     }
   } else if (type === 'nuxtjs') {
@@ -58,7 +69,8 @@ export default async (enableStatic: boolean, dir = process.cwd()): Promise<Confi
       configFile: fs.existsSync(nuxttsPath) ? nuxttsPath : undefined
     })
     const srcDir = path.posix.join(dir, config.srcDir ?? '')
-    const output = path.posix.join(srcDir, 'plugins')
+
+    output = output ?? path.posix.join(srcDir, 'plugins')
 
     if (!fs.existsSync(output)) fs.mkdirSync(output)
 
@@ -67,17 +79,23 @@ export default async (enableStatic: boolean, dir = process.cwd()): Promise<Confi
       input: path.posix.join(srcDir, 'pages'),
       staticDir: enableStatic ? path.posix.join(srcDir, 'static') : undefined,
       output,
+      ignorePath,
       trailingSlash: config.router?.trailingSlash,
       basepath: config.router?.base
     }
   } else {
     const srcDir = path.posix.join(dir, 'src')
 
+    output = output ?? path.join(srcDir, 'node_modules')
+
+    if (!fs.existsSync(output)) fs.mkdirSync(output)
+
     return {
       type,
       input: path.posix.join(srcDir, 'routes'),
       staticDir: enableStatic ? path.posix.join(dir, 'static') : undefined,
-      output: path.join(srcDir, 'node_modules')
+      output,
+      ignorePath
     }
   }
 }